+++ src/neurobrix/kernels/ops/dft.py
@@ -28,7 +28,7 @@ def dft_rfft_matrix_kernel(
     X_real = x @ cos.T and X_imag = x @ nsin.T.
     """
     pid = tl.program_id(0)
-    offs = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offs = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     total = N_bins * N
     mask = offs < total
     k = offs // N
@@ -55,7 +55,7 @@ def idft_c2r_matrix_kernel(
     real_out = X_real @ cos + X_imag @ sin.
     """
     pid = tl.program_id(0)
-    offs = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offs = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     total = N_bins * N
     mask = offs < total
     k = offs // N
