+++ src/neurobrix/kernels/ops/div.py
@@ -11,7 +11,7 @@ def div_forward_kernel(
 ):
     """out = x / y (tensor / tensor)"""
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     x = tl.load(x_ptr + offset, mask=mask)
@@ -27,7 +27,7 @@ def div_scalar_kernel(
 ):
     """out = x / scalar"""
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     x = tl.load(x_ptr + offset, mask=mask)
@@ -44,7 +44,7 @@ def div_scalar_dev_kernel(
     of div_scalar_kernel (host float(s) -> f32 arg == load -> f64 ->
     f32). Device-scalar increment 2026-08-15."""
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
     s = tl.load(s_ptr).to(tl.float64).to(tl.float32)
     x = tl.load(x_ptr + offset, mask=mask)
