+++ src/neurobrix/kernels/ops/flip_op.py
@@ -17,7 +17,7 @@ def flip_1d_kernel(
 ):
     """Reverse a 1D contiguous tensor: dst[i] = src[n_elements - 1 - i]."""
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     src_idx = n_elements - 1 - offset
@@ -43,7 +43,7 @@ def flip_strided_kernel(
     dim-axis index is mirrored.
     """
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     # Decompose flat index
