+++ src/neurobrix/kernels/ops/logical_xor.py
@@ -23,7 +23,7 @@ def logical_xor_kernel(
     Output is int8 (0 or 1).
     """
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     x = tl.load(x_ptr + offset, mask=mask, other=0.0)
