+++ src/neurobrix/kernels/ops/mse_loss.py
@@ -28,7 +28,7 @@ def mse_loss_partial_kernel(
     For reduction==2 (sum):  each block stores sum((inp-target)^2).
     """
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < M
 
     inp_val = tl.load(inp + offset, mask=mask, other=0).to(tl.float32)
@@ -68,7 +68,7 @@ def mse_loss_none_kernel(
 ):
     """Pointwise (no reduction): output[i] = (input[i] - target[i])^2."""
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     x = tl.load(inp_ptr + offset, mask=mask, other=0).to(tl.float32)
