+++ src/neurobrix/kernels/ops/prod.py
@@ -26,7 +26,7 @@ def prod_kernel_mid(
 ):
     """Pass 1: each program computes a partial product over its BLOCK_SIZE chunk."""
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     inp_ptrs = inp + offset
     mask = offset < M
     inp_val = tl.load(inp_ptrs, mask=mask, other=1.0).to(tl.float32)
